requirements.txt
pytest
pytest-playwright
pytest-json-report
python-docx